chore: add login confirmation page
Previously, the login process showed an ugly error message when the user
attempted to login prior the redirect to the backstroke api. This
change switches that screen to instead show a purpose-built screen that
notifies the user a redirect is taking place. It's a lot better in the
UX department, IMO.