code(extra-wide-cells): Render cells that have long titles as wider.
Previously, link items that had really long names would overflow
weirdly, often out of the bounds of the cell they were in. The solution
implemented here makes cells wider when they have a long name, such that
they won't overflow the bounds of their container. At 12 characters, the
cell will grow to be two wide, and at 32 characters, the cell will grow
to be three wide.

Given that we're using a non-monospace font, this isn't a perfect
solution.